refactor: extract semantic tool details
