+++ src/domain.rs
@@ -1,3 +1,4 @@
 pub(crate) mod activity;
 pub(crate) mod elicitation;
 pub(crate) mod session;
+pub(crate) mod tool;
